requirements.txt
streamlit
keras
Pillow
numpy
pandas
boto3